backend/requirements.txt
# Flask 框架
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# AI 服务
openai>=1.0.0
httpx>=0.25.0

# 文件处理
openpyxl>=3.1.0
python-pptx>=0.6.21
Pillow>=10.0.0

# 工具
uuid
